Removed factories folder as well as fixing all the appropriate namespaces
