+++ src/DotNetToolkit.Repository/Configuration/Options/IRepositoryOptions.cs
@@ -2,7 +2,6 @@
 {
     using Caching;
     using Conventions;
-    using Factories;
     using Interceptors;
     using Logging;
     using Mapper;
