+++ src/DotNetToolkit.Repository/Configuration/Options/Internal/RepositoryOptions.cs
@@ -2,7 +2,6 @@
 {
     using Caching;
     using Conventions;
-    using Factories;
     using Interceptors;
     using JetBrains.Annotations;
     using Logging;
