+++ src/DotNetToolkit.Repository/Internal/ConfigFile/IConfigurationSection.cs
@@ -1,10 +1,10 @@
 ﻿namespace DotNetToolkit.Repository.Internal.ConfigFile
 {
+    using Configuration;
     using Configuration.Caching;
     using Configuration.Interceptors;
     using Configuration.Logging;
     using Configuration.Mapper;
-    using Factories;
     using System;
     using System.Collections.Generic;
