+++ src/DotNetToolkit.Repository/Transactions/UnitOfWorkFactory.cs
@@ -1,15 +1,14 @@
-﻿namespace DotNetToolkit.Repository.Factories
+﻿namespace DotNetToolkit.Repository.Transactions
 {
     using Configuration.Options;
     using JetBrains.Annotations;
     using System;
-    using Transactions;
     using Utility;
 
     /// <summary>
     /// An implementation of <see cref="IUnitOfWorkFactory" />.
     /// </summary>
-    /// <seealso cref="DotNetToolkit.Repository.Factories.IUnitOfWorkFactory" />
+    /// <seealso cref="DotNetToolkit.Repository.Transactions.IUnitOfWorkFactory" />
     public class UnitOfWorkFactory : IUnitOfWorkFactory
     {
         #region Fields
