+++ test/DotNetToolkit.Repository.Integration.Test/DependencyInjectionTests.Microsoft.cs
@@ -4,7 +4,6 @@
     using Configuration.Options;
     using Data;
     using Extensions.Microsoft.DependencyInjection;
-    using Factories;
     using InMemory;
     using Microsoft.Extensions.DependencyInjection;
     using Services;
