+++ test/DotNetToolkit.Repository.Integration.Test/RepositoryTests.Update.cs
@@ -1,7 +1,6 @@
 ﻿namespace DotNetToolkit.Repository.Integration.Test
 {
     using Data;
-    using Factories;
     using System.Collections.Generic;
     using System.Threading.Tasks;
     using Xunit;
