+++ test/DotNetToolkit.Repository.Integration.Test/ServiceTests.GetDictionary.cs
@@ -1,8 +1,8 @@
 ﻿namespace DotNetToolkit.Repository.Integration.Test
 {
     using Data;
-    using Factories;
     using Queries;
+    using Services;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
