+++ test/DotNetToolkit.Repository.Integration.Test/ServiceTests.GetGroup.cs
@@ -1,8 +1,8 @@
 ﻿namespace DotNetToolkit.Repository.Integration.Test
 {
     using Data;
-    using Factories;
     using Queries;
+    using Services;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
