+++ test/DotNetToolkit.Repository.Integration.Test/ServiceTests.cs
@@ -1,7 +1,6 @@
 ﻿namespace DotNetToolkit.Repository.Integration.Test
 {
     using Data;
-    using Factories;
     using Services;
     using Xunit;
     using Xunit.Abstractions;
