+++ test/DotNetToolkit.Repository.Integration.Test/UnitOfWorkTests.cs
@@ -1,7 +1,6 @@
 ﻿namespace DotNetToolkit.Repository.Integration.Test
 {
     using Data;
-    using Factories;
     using System;
     using Transactions;
     using Xunit;
